changes from comments on the pull request v2
